Support for typeahead in Habitat site search
