+++ src/Feature/Catalog/code/App_Config/Include/Feature/Feature.Commerce.Catalog.config
@@ -66,7 +66,9 @@
         <solutionFramework>
             <indexing>
                 <providers>
-                    <add name="catalog" type="Sitecore.Feature.Commerce.Catalog.Infrastructure.Provider.IndexingProvider, Sitecore.Feature.Commerce.Catalog" resolve="true" patch:before="add[@name='fallback']" />
+                    <add name="productformatter" type="Sitecore.Feature.Commerce.Catalog.Infrastructure.Provider.ProductSearchResultFormatter, Sitecore.Feature.Commerce.Catalog" resolve="true" patch:before="add[@name='fallback']" />
+                    <add name="categoryformatter" type="Sitecore.Feature.Commerce.Catalog.Infrastructure.Provider.CategorySearchResultFormatter, Sitecore.Feature.Commerce.Catalog" resolve="true" patch:before="add[@name='fallback']" />
+                    <add name="catalog" type="Sitecore.Feature.Commerce.Catalog.Infrastructure.Provider.ProductCatalogQueryProvider, Sitecore.Feature.Commerce.Catalog" resolve="true" patch:before="add[@name='fallback']" />
                 </providers>
             </indexing>
         </solutionFramework>
