--- src/Project/Retail/code/Sitecore.Demo.Retail.csproj.sitecore
@@ -1,5 +1,0 @@
-﻿<Project>
-  <PropertyGroup>
-    <Site Server="retail.dev.local" UserName="sitecore\admin" IsRemoteSitecore="False" />
-  </PropertyGroup>
-</Project>
