+++ src/Project/Retail/serialization/Templates/Retail/Page Types/Article.yml
@@ -18,6 +18,9 @@ SharedFields:
     {BDD24F35-05E8-4466-8798-7D3DD6A6C991}
     {A1CBA309-D22B-46D5-80F8-2972C185363F}
     {8FD6C8B6-A9A4-4322-947E-90CE3D94916D}
+- ID: "93d1b217-b8f4-462e-babf-68298c9ce667"
+  Hint: __Boost
+  Value: 1.1
 - ID: "f7d48a55-2158-4f02-9356-756654404f73"
   Hint: __Standard values
   Value: "{450DD975-BB75-422E-BE67-79F50DA2143C}"
